requirements.txt
langchain
sentence-transformers
numpy
faiss-cpu
python-dotenv